+++ src/cli/subcommands/snapshot_cmd.rs
@@ -3,34 +3,35 @@
 
 use super::*;
 use crate::blocks::{tipset_keys_json::TipsetKeysJson, Tipset, TipsetKeys};
-use crate::car_backed_blockstore::{
-    self, CompressedCarV1BackedBlockstore, MaxFrameSizeExceeded, UncompressedCarV1BackedBlockstore,
-};
 use crate::chain::index::ChainIndex;
 use crate::cli::subcommands::{cli_error_and_die, handle_rpc_err};
 use crate::cli_shared::snapshot::{self, TrustedVendor};
 use crate::daemon::bundle::load_bundles;
+use crate::db::car::AnyCar;
 use crate::fil_cns::composition as cns;
 use crate::ipld::{recurse_links_hash, CidHashSet};
 use crate::networks::{calibnet, mainnet, ChainConfig, NetworkChain};
-use crate::rpc_api::{chain_api::ChainExportParams, progress_api::GetProgressType};
-use crate::rpc_client::{chain_ops::*, progress_ops::get_progress};
+use crate::rpc_api::chain_api::ChainExportParams;
+use crate::rpc_client::chain_ops::*;
 use crate::shim::machine::MultiEngine;
-use crate::utils::{io::ProgressBar, proofs_api::paramfetch::ensure_params_downloaded};
+use crate::utils::db::car_stream::CarStream;
+use crate::utils::proofs_api::paramfetch::ensure_params_downloaded;
 use anyhow::{bail, Context, Result};
 use chrono::Utc;
 use clap::Subcommand;
+use futures::TryStreamExt;
 use fvm_ipld_blockstore::Blockstore;
-use std::io::BufReader;
-use std::path::PathBuf;
+use human_repr::HumanCount;
+use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use tracing::info;
+use tempfile::NamedTempFile;
+use tokio::fs::File;
+use tokio::io::AsyncWriteExt;
 
 #[derive(Debug, Subcommand)]
 pub enum SnapshotCommands {
     /// Export a snapshot of the chain to `<output_path>`
     Export {
-        /// Snapshot output filename or directory. Defaults to
         /// `./forest_snapshot_{chain}_{year}-{month}-{day}_height_{epoch}.car.zst`.
         #[arg(short, default_value = ".", verbatim_doc_comment)]
         output_path: PathBuf,
@@ -108,54 +109,63 @@ impl SnapshotCommands {
                         chain_name,
                         Utc::now().date_naive(),
                         epoch,
+                        true,
                     )),
                     false => output_path.clone(),
                 };
 
+                let output_dir = output_path.parent().context("invalid output path")?;
+                let temp_path = NamedTempFile::new_in(output_dir)?.into_temp_path();
+
                 let params = ChainExportParams {
                     epoch,
                     recent_roots: config.chain.recent_state_roots,
-                    output_path,
+                    output_path: temp_path.to_path_buf(),
                     tipset_keys: TipsetKeysJson(chain_head.key().clone()),
                     skip_checksum,
                     dry_run,
                 };
 
-                let bar = Arc::new(tokio::sync::Mutex::new({
-                    let bar = ProgressBar::new(0);
-                    bar.message("Exporting snapshot | blocks ");
-                    bar
-                }));
-                tokio::spawn({
-                    let bar = bar.clone();
+                let handle = tokio::spawn({
+                    let tmp_file = temp_path.to_owned();
+                    let output_path = output_path.clone();
                     async move {
                         let mut interval =
-                            tokio::time::interval(tokio::time::Duration::from_secs(1));
+                            tokio::time::interval(tokio::time::Duration::from_secs_f32(0.25));
+                        println!("Getting ready to export...");
                         loop {
                             interval.tick().await;
-                            if let Ok((progress, total)) =
-                                get_progress((GetProgressType::SnapshotExport,), &None).await
-                            {
-                                let bar = bar.lock().await;
-                                if bar.is_finish() {
-                                    break;
-                                }
-                                bar.set_total(total);
-                                bar.set(progress);
-                            }
+                            let snapshot_size = std::fs::metadata(&tmp_file)
+                                .map(|meta| meta.len())
+                                .unwrap_or(0);
+                            print!(
+                                "{}{}",
+                                anes::MoveCursorToPreviousLine(1),
+                                anes::ClearLine::All
+                            );
+                            println!(
+                                "{}: {}",
+                                &output_path.to_string_lossy(),
+                                snapshot_size.human_count_bytes()
+                            );
+                            let _ = std::io::stdout().flush();
                         }
                     }
                 });
 
-                let out = chain_export(params, &config.client.rpc_token)
+                let hash_result = chain_export(params, &config.client.rpc_token)
                     .await
                     .map_err(handle_rpc_err)?;
 
-                bar.lock().await.finish_println(&format!(
-                    "Export completed. Snapshot located at {}",
-                    out.display()
-                ));
-                println!("\n");
+                handle.abort();
+                let _ = handle.await;
+
+                if let Some(hash) = hash_result {
+                    save_checksum(&output_path, hash).await?;
+                }
+                temp_path.persist(output_path)?;
+
+                println!("Export completed.");
                 Ok(())
             }
             Self::Fetch { directory, vendor } => {
@@ -173,82 +183,61 @@ impl SnapshotCommands {
                 check_stateroots,
                 snapshot,
             } => {
-                // this is all blocking...
-                use std::fs::File;
-                match CompressedCarV1BackedBlockstore::new(BufReader::new(File::open(&snapshot)?)) {
-                    Ok(store) => {
-                        validate_with_blockstore(
-                            store.roots(),
-                            Arc::new(store),
-                            check_links,
-                            check_network,
-                            check_stateroots,
-                        )
-                        .await
-                    }
-                    Err(error)
-                        if error.kind() == std::io::ErrorKind::Other
-                            && error.get_ref().is_some_and(|inner| {
-                                inner.downcast_ref::<MaxFrameSizeExceeded>().is_some()
-                            }) =>
-                    {
-                        bail!("The provided compressed car file cannot be used as a blockstore. Prepare it using `forest snapshot compress ...`")
-                    }
-                    Err(error) => {
-                        info!(%error, "file may be uncompressed, retrying as a plain CAR...");
-                        let store = UncompressedCarV1BackedBlockstore::new(File::open(&snapshot)?)?;
-                        validate_with_blockstore(
-                            store.roots(),
-                            Arc::new(store),
-                            check_links,
-                            check_network,
-                            check_stateroots,
-                        )
-                        .await
-                    }
-                }
+                let store = AnyCar::new(move || std::fs::File::open(&snapshot))?;
+                validate_with_blockstore(
+                    store.roots(),
+                    Arc::new(store),
+                    check_links,
+                    check_network,
+                    check_stateroots,
+                )
+                .await
             }
+
             Self::Compress {
                 source,
                 destination,
                 compression_level,
                 frame_size,
             } => {
-                // We've got a binary blob, and we're not exactly sure if it's compressed, and we can't just peek the header:
-                // For example, the zstsd magic bytes are a valid varint frame prefix:
-                assert_eq!(
-                    unsigned_varint::io::read_usize(&[0xFD, 0x2F, 0xB5, 0x28][..]).unwrap(),
-                    6141,
-                );
-                // so the best thing to do is to just try compressed and then uncompressed.
-                use car_backed_blockstore::zstd_compress_varint_manyframe;
-                use tokio::fs::File;
-                match zstd_compress_varint_manyframe(
-                    async_compression::tokio::bufread::ZstdDecoder::new(tokio::io::BufReader::new(
-                        File::open(&source).await?,
-                    )),
-                    File::create(&destination).await?,
+                let file = tokio::io::BufReader::new(File::open(&source).await?);
+                let mut block_stream = CarStream::new(file).await?;
+                let roots = std::mem::take(&mut block_stream.header.roots);
+
+                let mut dest = tokio::io::BufWriter::new(File::create(&destination).await?);
+
+                let frames = crate::db::car::forest::Encoder::compress_stream(
                     frame_size,
                     compression_level,
-                )
-                .await
-                {
-                    Ok(_num_frames) => Ok(()),
-                    Err(error) => {
-                        info!(%error, "file may be uncompressed, retrying as a plain CAR...");
-                        zstd_compress_varint_manyframe(
-                            File::open(&source).await?,
-                            File::create(&destination).await?,
-                            frame_size,
-                            compression_level,
-                        )
-                        .await?;
-                        Ok(())
-                    }
-                }
+                    block_stream.map_err(anyhow::Error::from),
+                );
+                crate::db::car::forest::Encoder::write(&mut dest, roots, frames).await?;
+                dest.flush().await?;
+                Ok(())
             }
         }
     }
+}
+
+/// Prints hex-encoded representation of SHA-256 checksum and saves it to a file
+/// with the same name but with a `.sha256sum` extension.
+async fn save_checksum(source: &Path, encoded_hash: String) -> Result<()> {
+    let checksum_file_content = format!(
+        "{encoded_hash} {}\n",
+        source
+            .file_name()
+            .and_then(std::ffi::OsStr::to_str)
+            .context("Failed to retrieve file name while saving checksum")?
+    );
+
+    let checksum_path = PathBuf::from(source).with_extension("sha256sum");
+
+    let mut checksum_file = tokio::fs::File::create(&checksum_path).await?;
+    checksum_file
+        .write_all(checksum_file_content.as_bytes())
+        .await?;
+    checksum_file.flush().await?;
+    Ok(())
 }
 
 // Check the validity of a snapshot by looking at IPLD links, the genesis block,